移植成功,待显示屏验证

+++ Seekfree_TC264_Opensource_Library/code/pid.c
@@ -4,9 +4,9 @@ void __initPID(PIDValue *pid,int32_t pCoef, int32_t iCoef, int32_t dCoef, int32_
     pid->pCorr = 0; pid->iCorr = 0; pid->dCorr = 0;
     pid->error[0] = 0; pid->error[1] = 0; pid->error[2] = 0;
     pid->errorInt = 0;
-    pid->errorIntMax = errorIntMax;
+    pid->errorIntMax = (float)errorIntMax;
     pid->pCoef = (float)pCoef; pid->iCoef = (float)iCoef; pid->dCoef = (float)dCoef;
-    pid->target = target;
+    pid->target = (float)target;
     pid->measurement = 0;
     pid->deltaOutput = 0;
 }
